Fix task turnpoint reordering. Some refactoring/code cleanup. Set back version name/code till ready for new release.
