+++ app/src/main/java/com/fisincorporated/soaringforecast/task/edit/DragOps.java
@@ -1,0 +1,5 @@
+package com.fisincorporated.soaringforecast.task.edit;
+
+public interface DragOps{
+    void dragCompleted();
+}
