+++ app/src/main/java/com/fisincorporated/soaringforecast/task/edit/TaskTurnpointsRecyclerViewAdapter.java
@@ -7,15 +7,17 @@ import android.view.ViewGroup;
 import com.fisincorporated.soaringforecast.R;
 import com.fisincorporated.soaringforecast.common.recycleradapter.GenericListClickListener;
 import com.fisincorporated.soaringforecast.common.recycleradapter.GenericRecyclerViewAdapter;
+import com.fisincorporated.soaringforecast.common.recycleradapter.GenericViewHolder;
 import com.fisincorporated.soaringforecast.databinding.TaskTurnpointView;
 import com.fisincorporated.soaringforecast.repository.TaskTurnpoint;
 import com.fisincorporated.soaringforecast.touchhelper.ItemTouchHelperAdapter;
+import com.fisincorporated.soaringforecast.touchhelper.ItemTouchHelperViewHolder;
 
 import java.util.Collections;
 
 public class TaskTurnpointsRecyclerViewAdapter
-        extends GenericRecyclerViewAdapter<TaskTurnpoint, TaskTurnpointViewHolder>
-        implements ItemTouchHelperAdapter {
+        extends GenericRecyclerViewAdapter<TaskTurnpoint, TaskTurnpointsRecyclerViewAdapter.TaskTurnpointViewHolder>
+        implements ItemTouchHelperAdapter, DragOps {
 
     private GenericListClickListener<TaskTurnpoint> itemClickListener;
     private TaskAndTurnpointsViewModel taskAndTurnpointsViewModel;
@@ -30,7 +32,7 @@ public class TaskTurnpointsRecyclerViewAdapter
     public TaskTurnpointViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         TaskTurnpointView binding = DataBindingUtil.inflate(
                 LayoutInflater.from(parent.getContext()), R.layout.task_turnpoint_detail, parent, false);
-        return new TaskTurnpointViewHolder(binding);
+        return new TaskTurnpointViewHolder(binding, this);
     }
 
     @Override
@@ -46,15 +48,55 @@ public class TaskTurnpointsRecyclerViewAdapter
         Collections.swap(getItems(), fromPosition, toPosition);
         notifyItemMoved(fromPosition, toPosition);
         taskAndTurnpointsViewModel.renumberTurnpoints();
-        notifyDataSetChanged();
         return true;
     }
+
+    @Override
+    public void dragCompleted() {
+        notifyDataSetChanged();
+    }
+
 
     @Override
     public void onItemDismiss(int position) {
         // Need to keep track of deleted turnpoints so can delete from db when(if) user saves changes
         taskAndTurnpointsViewModel.deleteTaskTurnpoint(position);
         notifyDataSetChanged();
+    }
+
+    public static class TaskTurnpointViewHolder extends GenericViewHolder<TaskTurnpoint, TaskTurnpointView>
+            implements ItemTouchHelperViewHolder {
+
+        private TaskTurnpointView viewDataBinding;
+        private DragOps dragOps;
+
+        TaskTurnpointViewHolder(TaskTurnpointView bindingView, DragOps dragOps) {
+            super(bindingView);
+            this.dragOps = dragOps;
+            viewDataBinding = bindingView;
+        }
+
+        public void onBind(TaskTurnpoint item, int position) {
+            viewDataBinding.setTaskTurnpoint(item);
+            viewDataBinding.setPosition(position);
+        }
+
+        @Override
+        public TaskTurnpointView getViewDataBinding() {
+            return viewDataBinding;
+        }
+
+        @Override
+        public void onItemSelected() {
+            itemView.setBackgroundColor(itemView.getResources().getColor(R.color.drag_color));
+        }
+
+        @Override
+        public void onItemClear() {
+            itemView.setBackgroundColor(itemView.getResources().getColor(R.color.drag_drop));
+            dragOps.dragCompleted();
+        }
+
     }
 
 }
