+++ app/src/main/java/com/fisincorporated/soaringforecast/task/list/TaskListViewHolder.java
@@ -1,11 +1,13 @@
 package com.fisincorporated.soaringforecast.task.list;
 
 
+import com.fisincorporated.soaringforecast.R;
 import com.fisincorporated.soaringforecast.common.recycleradapter.GenericViewHolder;
 import com.fisincorporated.soaringforecast.databinding.TaskView;
 import com.fisincorporated.soaringforecast.repository.Task;
+import com.fisincorporated.soaringforecast.touchhelper.ItemTouchHelperViewHolder;
 
-public class TaskListViewHolder extends GenericViewHolder<Task, TaskView> {
+public class TaskListViewHolder extends GenericViewHolder<Task, TaskView> implements ItemTouchHelperViewHolder {
 
     private TaskView viewDataBinding;
 
@@ -22,6 +24,16 @@ public class TaskListViewHolder extends GenericViewHolder<Task, TaskView> {
     @Override
     public TaskView getViewDataBinding() {
         return viewDataBinding;
+    }
+
+    @Override
+    public void onItemSelected() {
+       itemView.setBackgroundColor(itemView.getResources().getColor(R.color.drag_color));
+    }
+
+    @Override
+    public void onItemClear() {
+       itemView.setBackgroundColor(itemView.getResources().getColor(R.color.drag_drop));
     }
 
 }
